finish loader and fix some bugs
