+++ Assets/Shaders/fragment_shader.glsl
@@ -4,6 +4,7 @@ in vec3 vertexColor;
 out vec4 FragColor;
 
 void main()"
+void main()
 {
  FragColor = vec4(vertexColor.r, vertexColor.g, vertexColor.b, 1.0f);
 };
