+++ Assets/Shaders/vertex_shader.glsl
@@ -5,13 +5,17 @@ layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aColor;
 
 uniform mat4 worldMatrix;
+/*uniform mat4 worldMatrix; modelViewProjection 
 uniform mat4 viewMatrix = mat4(1.0); // default value for view matrix (identity)
-uniform mat4 projectionMatrix = mat4(1.0);
+uniform mat4 projectionMatrix = mat4(1.0);*/
 
 out vec3 vertexColor;
 void main()"
+void main()
 {
   vertexColor = aColor;
   mat4 modelViewProjection = projectionMatrix * viewMatrix * worldMatrix;
   gl_Position = modelViewProjection * vec4(aPos.x, aPos.y, aPos.z, 1.0);
+  //mat4 modelViewProjection = projectionMatrix * viewMatrix * worldMatrix;
+  gl_Position =  vec4(aPos.x, aPos.y, aPos.z, 1.0);
 };
