+++ Source/cube.h
@@ -1,0 +1,7 @@
+#pragma once
+class cube {
+public:
+	cube();
+	~cube();
+	int createCubeArray();
+};
