+++ Source/grids.h
@@ -1,0 +1,7 @@
+#pragma once
+class grid {
+public:
+	grid();
+	~grid();
+	int createGridArray();
+};
