--- Source/lab01.cpp
@@ -1,249 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 20/06/2019.
-//
-// Inspired by the following tutorials:
-// - https://learnopengl.com/Getting-started/Hello-Window
-// - https://learnopengl.com/Getting-started/Hello-Triangle
-
-#include <iostream>
-
-
-#define GLEW_STATIC 1   // This allows linking with Static Library on Windows, without DLL
-#include <GL/glew.h>    // Include GLEW - OpenGL Extension Wrangler
-
-#include <GLFW/glfw3.h> // GLFW provides a cross-platform interface for creating a graphical context,
-                        // initializing OpenGL and binding inputs
-
-#include <glm/glm.hpp>  // GLM is an optimized math library with syntax to similar to OpenGL Shading Language
-
-
-
-const char* getVertexShaderSource()
-{
-	// Insert Shaders here ...
-	// For now, you use a string for your shader code, in the assignment, shaders will be stored in .glsl files
-
-	return
-		"#version 330 core\n"
-		"layout (location = 0) in vec3 aPos;" // input to the shader program
-		"layout (location = 1) in vec3 aColor;"
-		"out vec3 vertexColor;"  // output to the shader program
-		"void main()"
-		"{"
-		"   vertexColor = aColor;"
-		"   gl_Position = vec4(aPos.x, aPos.y, aPos.z, 1.0);" //homogenous coordinates = vec3 + 0 or 1 . vec3 +0 = vector and vec3 +1 = position
-		"}";
-}
-
-const char* getFragmentShaderSource()
-{
-	return
-		"#version 330 core\n"
-		"in vec3 vertexColor;"
-		"out vec4 FragColor;"
-		"void main()"
-		"{"
-		"   FragColor = vec4(vertexColor.r, vertexColor.g, vertexColor.b, 1.0f);" // rgba
-		"}";
-}
-
-
-int compileAndLinkShaders()
-{
-	// compile and link shader program
-	// return shader program id
-	// ------------------------------------
-
-
-	// vertex shader
-	int vertexShader = glCreateShader(GL_VERTEX_SHADER);
-	const char* vertexShaderSource = getVertexShaderSource();
-	glShaderSource(vertexShader, 1, &vertexShaderSource, NULL);
-	glCompileShader(vertexShader);
-
-	// check for shader compile errors
-	int success;
-	char infoLog[512];
-	glGetShaderiv(vertexShader, GL_COMPILE_STATUS, &success);
-	if (!success)
-	{
-		glGetShaderInfoLog(vertexShader, 512, NULL, infoLog);
-		std::cerr << "ERROR::SHADER::VERTEX::COMPILATION_FAILED\n" << infoLog << std::endl;
-	}
-
-	// fragment shader
-	int fragmentShader = glCreateShader(GL_FRAGMENT_SHADER);
-	const char* fragmentShaderSource = getFragmentShaderSource();
-	glShaderSource(fragmentShader, 1, &fragmentShaderSource, NULL);
-	glCompileShader(fragmentShader);
-
-	// check for shader compile errors
-	glGetShaderiv(fragmentShader, GL_COMPILE_STATUS, &success);
-	if (!success)
-	{
-		glGetShaderInfoLog(fragmentShader, 512, NULL, infoLog);
-		std::cerr << "ERROR::SHADER::FRAGMENT::COMPILATION_FAILED\n" << infoLog << std::endl;
-	}
-
-	// link shaders
-	int shaderProgram = glCreateProgram();
-	glAttachShader(shaderProgram, vertexShader);
-	glAttachShader(shaderProgram, fragmentShader);
-	glLinkProgram(shaderProgram);
-
-	// check for linking errors
-	glGetProgramiv(shaderProgram, GL_LINK_STATUS, &success);
-	if (!success) {
-		glGetProgramInfoLog(shaderProgram, 512, NULL, infoLog);
-		std::cerr << "ERROR::SHADER::PROGRAM::LINKING_FAILED\n" << infoLog << std::endl;
-	}
-
-	glDeleteShader(vertexShader);
-	glDeleteShader(fragmentShader);
-
-	return shaderProgram;
-}
-
-int createVertexArrayObject()
-{
-	// A vertex is a point on a polygon, it contains positions and other data (eg: colors)
-	glm::vec3 vertexArray[] = {
-		glm::vec3(0.0f,  0.5f, 0.0f),  // top center position in Normalize coordinate vertices
-		glm::vec3(1.0f,  0.0f, 0.0f),  // top center color (red)
-		glm::vec3(0.5f, -0.5f, 0.0f),  // bottom right
-		glm::vec3(0.0f,  1.0f, 0.0f),  // bottom right color (green)
-		glm::vec3(-0.5f, -0.5f, 0.0f),  // bottom left
-		glm::vec3(0.0f,  0.0f, 1.0f),  // bottom left color (blue)
-	};
-
-	// Create a vertex array
-	GLuint vertexArrayObject;
-	glGenVertexArrays(1, &vertexArrayObject);
-	glBindVertexArray(vertexArrayObject);
-
-
-	// Upload Vertex Buffer to the GPU, keep a reference to it (vertexBufferObject)
-	GLuint vertexBufferObject;  // object ID
-	glGenBuffers(1, &vertexBufferObject);
-	glBindBuffer(GL_ARRAY_BUFFER, vertexBufferObject); // create a vao and assign its id to the variable
-	glBufferData(GL_ARRAY_BUFFER, sizeof(vertexArray), vertexArray, GL_STATIC_DRAW); // binding a vao, this tells opengl that we want to use this for the following code
-
-	// prase the data we sent to buffer and assign positon id 0 and color 1
-	glVertexAttribPointer(0,                   // attribute 0 matches aPos in Vertex Shader (position id in shader)
-		3,                   // size
-		GL_FLOAT,            // type
-		GL_FALSE,            // normalized?
-		2 * sizeof(glm::vec3), // stride - each vertex contain 2 vec3 (position, color)
-		(void*)0             // array buffer offset
-	);
-	glEnableVertexAttribArray(0);
-
-
-	glVertexAttribPointer(1,                            // attribute 1 matches aColor in Vertex Shader
-		3,
-		GL_FLOAT,
-		GL_FALSE,
-		2 * sizeof(glm::vec3),
-		(void*)sizeof(glm::vec3)      // color is offseted a vec3 (comes after position)
-	);
-	glEnableVertexAttribArray(1);
-
-	glBindBuffer(GL_ARRAY_BUFFER, 0); // VAO already stored the state we just defined, safe to unbind buffer
-	// tells the opengl we dont use the vbo anymore
-	glBindVertexArray(0); // Unbind to not modify the VAO
-	 
-	return vertexArrayObject; //VAO contains the VBO contains the vertex data
-}
-
-
-int main(int argc, char*argv[])
-{
-    // Initialize GLFW and OpenGL version
-    glfwInit();
-
-#if defined(PLATFORM_OSX)	
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3); 
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2); 
-	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-    glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
-#else
-	// On windows, we set OpenGL version to 2.1, to support more hardware
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 2);
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 1);
-	//glfwWindowHint(GLFW_OPENGL_PROFILE,GLFW_OPENGL_CORE_PROFILE );
-
-#endif
-
-    // Create Window and rendering context using GLFW, resolution is 800x600
-    GLFWwindow* window = glfwCreateWindow(800, 600, "Comp371 - Lab 01", NULL, NULL);
-    if (window == NULL)
-    {
-        std::cerr << "Failed to create GLFW window" << std::endl;
-        glfwTerminate();
-        return -1;
-    }
-    glfwMakeContextCurrent(window);
-    
-
-    // Initialize GLEW
-    glewExperimental = true; // Needed for core profile
-    if (glewInit() != GLEW_OK) {
-        std::cerr << "Failed to create GLEW" << std::endl;
-        glfwTerminate();
-        return -1;
-    }
-
-    // Black background
-    glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
-    // rgb (0-255) value / 255
-    
-    // Compile and link shaders here ...
-    int shaderProgram = compileAndLinkShaders();
-    
-    // Define and upload geometry to the GPU here ...
-    int vao = createVertexArrayObject();
-    
-    
-    // Entering Main Loop
-    while(!glfwWindowShouldClose(window))
-    {
-        // Each frame, reset color of each pixel to glClearColor
-        glClear(GL_COLOR_BUFFER_BIT);
-        
-		//Updata our resources
-        // Draw the objects
-        // TODO - draw rainbow triangle
-		glUseProgram(shaderProgram);
-		glBindVertexArray(vao);
-		glDrawArrays(GL_TRIANGLES, 0, 3); // 3 vertices, starting at vertex array index 0
-		glBindVertexArray(0);
-		//glBindVertexArray(vao1);
-		//
-		// procedure for using VAO and VBO
-		// before the game loop, initialize the vaos:
-		// 1. generate, store and bind a vao
-		// 2. generate, store and bind the vbo(s)  xyzxyz rgbrgb stst
-		// 3. unbind the vbo and the vaos
-		// in the main loop, render:
-		// bind vao
-		// glDraw*()
-		// unbind the vao
-
-
-        // End frame
-        glfwSwapBuffers(window); //render in the background buffer, and display the front buffer withn rendered image
-        
-        // Detect inputs
-        glfwPollEvents();
-        
-        if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS)
-            glfwSetWindowShouldClose(window, true);
-    }
-    
-    // Shutdown GLFW
-    glfwTerminate();
-    
-	return 0;
-}
