+++ project/Labs.log
@@ -1,4 +1,3 @@
 ﻿LINK : warning LNK4098: defaultlib 'LIBCMT' conflicts with use of other libs; use /NODEFAULTLIB:library
-LINK : warning LNK4098: defaultlib 'MSVCRT' conflicts with use of other libs; use /NODEFAULTLIB:library
-glew32s.lib(glew.obj) : warning LNK4099: PDB 'vc120.pdb' was not found with 'glew32s.lib(glew.obj)' or at 'I:\university\summer2020\comp371\project\ComputerGraphics\Bin\vc120.pdb'; linking object as if no debug info
+﻿  main.cpp
   Labs.vcxproj -> I:\university\summer2020\comp371\project\ComputerGraphics\project\/../Bin/Labs-Debug.exe
