+++ project/Labs.sln
@@ -1,20 +1,31 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.960
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Labs", "Labs.vcxproj", "{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Win32 = Debug|Win32
+		Debug|x64 = Debug|x64
 		Release|Win32 = Release|Win32
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|Win32.ActiveCfg = Debug|Win32
 		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|Win32.Build.0 = Debug|Win32
+		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|x64.ActiveCfg = Debug|x64
+		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|x64.Build.0 = Debug|x64
 		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|Win32.ActiveCfg = Release|Win32
 		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|Win32.Build.0 = Release|Win32
+		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|x64.ActiveCfg = Release|x64
+		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {43CFF9A4-9404-4F67-A68F-DBAFCE3DD969}
 	EndGlobalSection
 EndGlobal
