--- project/cube.h
@@ -1,6 +1,0 @@
-#pragma once
-class cube {
-public:
-	cube();
-	int createCubeArray();
-};
