--- project/grids.h
@@ -1,6 +1,0 @@
-#pragma once
-class grid {
-public:
-	grid();
-	int createGridArray();
-};
