+++ project/x64/Debug/Labs.tlog/Labs.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.16299.0
+Debug|x64|I:\university\summer2020\comp371\project\ComputerGraphics\project\|
